requirements.txt
pydantic_ai_bedrock
pydantic_ai
python-dotenv
tomli
pyyaml
